refactor(retention): isolate policy and account deletion adapters
