+++ apps/web/src/lib/server/account-binding.ts
@@ -1,8 +1,6 @@
 import { type AccountLifecycleBinding } from '@b2b-saas-starter/capabilities/governance/account-lifecycle'
-import * as schema from '@b2b-saas-starter/db/schema'
+import { makeAdminAccountLifecycleBinding } from '@b2b-saas-starter/capabilities/governance/account-lifecycle-admin.live'
 import { env } from 'cloudflare:workers'
-import { drizzle } from 'drizzle-orm/d1'
-import { and, eq } from 'drizzle-orm'
 
 import { sessionCall } from './plugin-call'
 
@@ -46,45 +44,5 @@ export const webAccountLifecycleBinding: AccountLifecycleBinding = {
   }
 }
 
-/**
- * The admin remove-user endpoint has an admin session, not the target user's
- * session. Its lifecycle hook therefore cannot use the session-bound
- * organization endpoints above without mutating the wrong account. The admin
- * endpoint is already Better Auth's privileged, role-checked surface, so its
- * teardown uses the same D1 rows directly and leaves the capability in charge
- * of the ownership rule and audit records.
- */
-export const webAdminAccountLifecycleBinding: AccountLifecycleBinding = {
-  leaveWorkspace: async (input) => {
-    const db = adminDeletionDatabase()
-    await db
-      .delete(schema.workspaceMembers)
-      .where(
-        and(
-          eq(schema.workspaceMembers.id, input.memberId),
-          eq(schema.workspaceMembers.workspaceId, input.workspaceId)
-        )
-      )
-  },
-  deleteWorkspace: async (input) => {
-    const db = adminDeletionDatabase()
-    await db
-      .delete(schema.workspaces)
-      .where(eq(schema.workspaces.id, input.workspaceId))
-  },
-  deleteUser: () => {
-    // The admin endpoint owns the user-row delete. This binding only supplies
-    // the pre-delete workspace teardown, so reaching this method is a wiring
-    // error rather than a second deletion path.
-    // oxlint-disable-next-line effect/noNewPromise, effect/noNewError -- this promise-shaped binding is an impossible wiring path; the Better Auth admin endpoint owns the row delete.
-    return Promise.reject(new Error('admin account binding cannot delete the user row'))
-  }
-}
-
-function adminDeletionDatabase() {
-  if (env.DB === undefined) {
-    // oxlint-disable-next-line effect/noThrowStatement, effect/noNewError -- a deployed admin deletion cannot proceed without its required D1 binding.
-    throw new Error('admin account deletion requires the D1 binding')
-  }
-  return drizzle(env.DB)
-}
+/** The admin endpoint's authorized target-account teardown, owned by governance. */
+export const webAdminAccountLifecycleBinding = makeAdminAccountLifecycleBinding(env.DB)
